[리펙토링][3,4단게] 테스트 패키지 구조 변경
